refactor(cex): make CexHarness and BmcTraceWrapper templates
